chore: Added bootstrap script for windows, upgraded premake version

+++ sandbox/SandboxApp/src/SandboxApp.cpp
@@ -27,7 +27,7 @@ public:
 	{
 
 	}
-	void OnUpdate(double timestep) override
+	void OnUpdate(float timestep) override
 	{
 		Xen::RenderCommand::Clear();
 		Xen::RenderCommand::SetClearColor({ color[0], color[1], color[2], color[3] });
